pyproject.toml
[tool.poetry]
name = "simple_calendar_drf"
version = "0.1.0"
description = ""
authors = ["Andrey Smakotin <slamfromm@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.3"
djangorestframework = "^3.13.1"
celery = "^5.2.3"
djoser = "^2.1.0"
django-environ = "^0.8.1"
psycopg2 = "^2.9.3"
ics = "^0.7"
beautifulsoup4 = "^4.10.0"
lxml = "^4.8.0"
django-filter = "^21.1"
pika = "^1.2.0"
rabbitmq = "^0.2.0"
gevent = "^21.12.0"
django-debug-toolbar = "^3.2.4"
tqdm = "^4.64.0"
flower = "^1.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
